Further tweaking and refinement. Move all remote fetching logic into the factory - remove none web accessible public methods from the controller. xibosignage/xibo#1320
