+++ views/dataset-form-add.twig
@@ -40,7 +40,7 @@
         <div class="col-md-12">
             <ul class="nav nav-tabs" role="tablist">
                 <li class="active"><a href="#general" role="tab" data-toggle="tab"><span>{% trans "General" %}</span></a></li>
-                <li><a href="#gateway" role="tab" data-toggle="tab"><span>{% trans "Data" %}</span></a></li>
+                <li><a href="#gateway" role="tab" data-toggle="tab"><span>{% trans "Remote" %}</span></a></li>
                 <li><a href="#auth" role="tab" data-toggle="tab"><span>{% trans "Authentication" %}</span></a></li>
                 <li><a href="#data" role="tab" data-toggle="tab"><span>{% trans "Data" %}</span></a></li>
                 <li><a href="#params" role="tab" data-toggle="tab"><span>{% trans "Advanced" %}</span></a></li>
