+++ views/dataset-page.twig
@@ -186,7 +186,7 @@
         table.on('processing.dt', dataTableProcessing);
 
         function XiboDatasetRemoteTestRequest($form) {
-            XiboRemoteRequest("{{ urlFor("dataSet.test.remote") }}", $("#dataSetRemoteForm").serializeObject(), function(response) {
+            XiboRemoteRequest("{{ urlFor("dataSet.test.remote") }}", $form.serializeObject(), function(response) {
                 $("#datasetRemoteTestRequestResult").html('<pre>' + JSON.stringify(response.data, null, 3) + '</pre>');
             });
         }
